feat: 결제 금액의 10%를 포인트로 지급하는 기능 추가

+++ src/main/java/cart/application/OrderService.java
@@ -1,13 +1,16 @@
 package cart.application;
 
 import cart.dao.CartItemDao;
+import cart.dao.PointDao;
 import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Order;
 import cart.domain.OrderItem;
+import cart.domain.PointPolicy;
 import cart.domain.ShippingPolicy;
 import cart.dto.request.OrderItemDto;
 import cart.dto.request.OrderRequest;
+import cart.dto.response.OrderAdditionResponse;
 import cart.dto.response.OrderDetailsDto;
 import cart.dto.response.OrderResponse;
 import cart.dto.response.ProductResponse;
@@ -15,6 +18,8 @@ import cart.exception.CartItemException.IllegalMember;
 import cart.exception.CartItemException.InvalidCartItem;
 import cart.exception.CartItemException.QuantityNotSame;
 import cart.exception.CartItemException.UnknownCartItem;
+import cart.exception.OrderException;
+import cart.exception.OrderException.LackOfPoint;
 import cart.repository.OrderRepository;
 import cart.repository.ShippingPolicyRepository;
 import org.springframework.stereotype.Service;
@@ -32,18 +37,21 @@ public class OrderService {
     private final CartItemDao cartItemDao;
     private final OrderRepository orderRepository;
     private final ShippingPolicyRepository shippingPolicyRepository;
+    private final PointDao pointDao;
 
     public OrderService(
             final CartItemDao cartItemDao,
             final OrderRepository orderRepository,
-            final ShippingPolicyRepository shippingPolicyRepository
+            final ShippingPolicyRepository shippingPolicyRepository,
+            final PointDao pointDao
     ) {
         this.cartItemDao = cartItemDao;
         this.orderRepository = orderRepository;
         this.shippingPolicyRepository = shippingPolicyRepository;
+        this.pointDao = pointDao;
     }
 
-    public long saveOrder(final Member member, final OrderRequest orderRequest) {
+    public OrderAdditionResponse saveOrder(final Member member, final OrderRequest orderRequest) {
         List<CartItem> cartItems = findCartItems(orderRequest);
 
         checkIllegalMember(member, cartItems);
@@ -52,15 +60,24 @@ public class OrderService {
 
         ShippingPolicy shippingPolicy = shippingPolicyRepository.findShippingPolicy();
         List<OrderItem> orderItems = OrderItem.of(cartItems);
-        Order order = Order.of(member, shippingPolicy.calculateShippingFee(orderItems), orderItems);
+        Order order = Order.of(member, shippingPolicy.calculateShippingFee(orderItems), orderRequest.getUsedPoint() ,orderItems);
         order.checkTotalProductsPrice(orderRequest.getTotalProductsPrice());
         order.checkShippingFee(orderRequest.getShippingFee());
+
+        long totalPoint = pointDao.selectByMemberId(member.getId());
+        long usedPoint = orderRequest.getUsedPoint();
+        if(totalPoint < usedPoint){
+            throw new LackOfPoint();
+        }
+        long newEarnedPoint = PointPolicy.getEarnedPoint(order.getPayment());
 
         long orderId = orderRepository.save(order);
         for (CartItem cartItem : cartItems) {
             cartItemDao.deleteById(cartItem.getId());
         }
-        return orderId;
+        pointDao.update(member.getId(), totalPoint - usedPoint + newEarnedPoint);
+
+        return new OrderAdditionResponse(orderId, newEarnedPoint);
     }
 
     private void checkQuantity(final List<OrderItemDto> orders, final List<CartItem> cartItems) {
@@ -119,7 +136,7 @@ public class OrderService {
                 .stream()
                 .map(item -> new OrderDetailsDto(item.getQuantity(), ProductResponse.of(item.getProduct()))).
                 collect(Collectors.toUnmodifiableList());
-        return new OrderResponse(orderId, order.getCreatedAt(), order.getTotalProductsPrice(), order.getShippingFee(), orderDetails);
+        return new OrderResponse(orderId, order.getCreatedAt(), order.getTotalProductsPrice(), order.getShippingFee(), order.getUsedPoint(), orderDetails);
     }
 
     public List<OrderResponse> getOrdersByMember(final Member member) {
@@ -130,7 +147,7 @@ public class OrderService {
                     .stream()
                     .map(item -> new OrderDetailsDto(item.getQuantity(), ProductResponse.of(item.getProduct()))).
                     collect(Collectors.toUnmodifiableList());
-            orderResponses.add(new OrderResponse(order.getId(), order.getCreatedAt(), order.getTotalProductsPrice(), order.getShippingFee(), orderDetails));
+            orderResponses.add(new OrderResponse(order.getId(), order.getCreatedAt(), order.getTotalProductsPrice(), order.getShippingFee(), order.getUsedPoint(), orderDetails));
         }
         return orderResponses;
     }
